requirements.txt
# Django 6.x, Python >= 3.12
Django>=6.0,<7.0
psycopg[binary]>=3.2
python-dotenv>=1.0
Pillow>=10.0

# Продакшен-сервер
gunicorn>=21.2

# Тесты
pytest-django>=4.9
pytest-cov>=6.0

# reCAPTCHA для защиты форм
django-recaptcha>=4.0
